requirements.txt
aiogram>=3.5,<4
aiohttp>=3.9
beautifulsoup4>=4.12
python-dotenv>=1.0
